requirements.txt
fastapi
uvicorn
streamlit
mysql-connector-python
python-dotenv
requests
pydantic